// src/api/models/userRegisterRequest.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * User registration request
 */
export interface UserRegisterRequest {
  /**
   * Username (3-50 characters)
   * @minLength 3
   * @maxLength 50
   */
  username: string;
  /**
   * Password (6-20 characters)
   * @minLength 6
   * @maxLength 2147483647
   */
  password: string;
  /** Email address */
  email: string;
  phone?: string;
}